feat!: replace plugin with LoggerInterface implementation
BREAKING CHANGE: We no longer plugin around the query processor (which is a good thing). However, Magento's out-of-box log data doesn't give the exact query sent. This is likely due to the fact that it is possible for mutation data like user passwords to appear in the log if developers fail to use variables appropriately. However, I'm not going to compensate for naive developers. To keep existing functionality, we've added the full query (not variables) to the log data when this package is enabled. I would avoid using this in production environments if at all possible if you're concerned about the quality of your development team.